Move topic-related decoding to own data location
- Move abi/decode/event to topic/decode
- Move abi/decode/abi to abi/decode
- Export Topic in index
