+++ packages/codec/lib/decode.ts
@@ -11,6 +11,7 @@ import * as Memory from "@truffle/codec/memory";
 import * as Special from "@truffle/codec/special";
 import * as Stack from "@truffle/codec/stack";
 import * as Storage from "@truffle/codec/storage";
+import * as Topic from "@truffle/codec/topic";
 
 export default function* decode(
   dataType: Format.Types.Type,
@@ -42,7 +43,7 @@ export default function* decode(
       return yield* Abi.Decode.decodeAbi(dataType, pointer, info, options);
 
     case "eventtopic":
-      return yield* Abi.Decode.decodeTopic(dataType, pointer, info, options);
+      return yield* Topic.Decode.decodeTopic(dataType, pointer, info, options);
 
     case "memory":
       //NOTE: this case should never actually occur, but I'm including it
