+++ packages/codec/lib/topic/index.ts
@@ -1,0 +1,2 @@
+import * as Decode from "./decode";
+export { Decode };
